Corresponding Source Code to ISC
This is the latest version if the ISC source code for people who are
interested in working on this project. The origional developer sold many
of his apps including the wonderful ISC and the project was abandoned.
Thankfully, the GPL does not allow revoking the rights included in it,
so I am distributing this source code under such terms of the GPLv2.
Fork away. This commit is for ISC version v6.